Add warnings about Legacy auth and Reader/Writer IAM
When using Reader or Writer IAM roles in a service credential, one must
be careful to also be using the _Use IAM Only_ option on account
provisioning -- otherwise the service credential contains an
IAM Manager role level Legacy username and password.

Documenting this is perhaps not enough, but at least we can do that
to start with.