Create responsibility document.
We are required to create a service-specific roles and responsibilities doc:
Instructions: see UX031 above for details
Service OMs must review the core platform documentation on Shared responsibilities for IBM Cloud offerings and determine if there is a need for a service-specific documentation topic that provides additional information on the required responsibilities.
If you require an additional topic in your service's documentation, fill in the documentation template which explains the differences in responsibilities in different areas and links to the core platform
 documentation. Make a copy of the template in an editor of your choice and add the required information.
The service OM then passes the responsibility of publishing the documentation to the team's content (docs) focal who will then follow the guidance to publish the documentation in the required section of the product doc.

https://test.cloud.ibm.com/docs/developing/writing?topic=writing-responsibilities-content-guidance

https://github.ibm.com/Bluemix/docs/blob/staging/developing/content-kit/responsibilities-template.md