cbr: address feedback
Adds a note on IAM token expiry. When IAM IP allowlists are changed it
may take up to 60 minutes for the changes to take effect since
IP allowlisting is only enforced at token creation time.

This also replaces usage of `access list` with `IP allowlist` given
that's what IAM call it, and fixes a couple of grammar issues.